streamline api towards vulkan
